fix type

+++ tests/tcg/plugins/discons.c
@@ -53,8 +53,8 @@ static void insn_exec(unsigned int vcpu_index, void *userdata)
 
 static void vcpu_tb_trans(qemu_plugin_id_t id, struct qemu_plugin_tb *tb)
 {
-    size_t i;
-    size_t n_insns = qemu_plugin_tb_n_insns(tb);
+    uint64_t i;
+    uint64_t n_insns = qemu_plugin_tb_n_insns(tb);
 
     g_autoptr(GString) s = g_string_new("");
     uint64_t vaddr = qemu_plugin_tb_vaddr(tb);
